latest commit

+++ src/main/java/com/ascentdev/wims/repository/RefULDRepository.java
@@ -14,7 +14,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
  * @author
  * ASCENT
  */
-public interface RefULDRepository extends JpaRepository<RefULDEntity, Long>{
+public interface RefULDRepository extends JpaRepository<RefULDEntity, Integer>{
   
   List<RefULDEntity> findByFlightNumber(String flightNumber);
   RefULDEntity findByUldNo(String uld_no);
